refactor ALS into a package
