+++ lenskit/algorithms/als/__init__.py
@@ -1,0 +1,8 @@
+"""
+LensKit ALS implementations.
+"""
+
+from .explicit import BiasedMF
+from .implicit import ImplicitMF
+
+__all__ = ["BiasedMF", "ImplicitMF"]
